requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
drf-yasg
gunicorn
django-heroku
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
python-decouple==3.4